Move accordion to modules
