+++ next/components/HomePage/SectionFaq.tsx
@@ -2,7 +2,7 @@ import Arrow from '@assets/images/arrow-right.svg'
 import { useUIContext } from '@bratislava/common-frontend-ui-context'
 import { ComponentHomepageFaqSection } from '@bratislava/strapi-sdk-city-library'
 import { CallToAction } from '@bratislava/ui-city-library'
-import Accordion from '@components/ui/Accordion/Accordion'
+import Accordion from '@modules/common/Accordion'
 import MLink from '@modules/common/MLink'
 import { useTranslation } from 'next-i18next'
 import React from 'react'
