+++ next/components/ui/index.ts
@@ -1,4 +1,3 @@
-export * from './Accordion/Accordion'
 export * from './ArticleCard/ArticleCard'
 export * from './Banner/Banner'
 export * from './BookNewsDetail/BookNewsDetail'
@@ -43,3 +42,4 @@ export * from './TableRow/TableRow'
 export * from './TextArea/TextArea'
 export * from './Upload/Upload'
 export * from './Video/Video'
+export * from '@modules/common/Accordion'
